APIs/twilio.com/twilio_studio_v2/1.55.0/python-aiohttp-client/tests/test_studio_v2_flow_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.flow_enum_status import FlowEnumStatus
from openapi_server.models.list_flow_response import ListFlowResponse
from openapi_server.models.studio_v2_flow import StudioV2Flow


pytestmark = pytest.mark.asyncio

@pytest.mark.skip("application/x-www-form-urlencoded not supported by Connexion")
async def test_create_flow(client):
    """Test case for create_flow

    
    """
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/x-www-form-urlencoded',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    data = {
        'commit_message': 'commit_message_example',
        'definition': None,
        'friendly_name': 'friendly_name_example',
        'status': openapi_server.FlowEnumStatus()
        }
    response = await client.request(
        method='POST',
        path='/v2/Flows',
        headers=headers,
        data=data,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_delete_flow(client):
    """Test case for delete_flow

    
    """
    headers = { 
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    response = await client.request(
        method='DELETE',
        path='/v2/Flows/{sid}'.format(sid='sid_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_fetch_flow(client):
    """Test case for fetch_flow

    
    """
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    response = await client.request(
        method='GET',
        path='/v2/Flows/{sid}'.format(sid='sid_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_list_flow(client):
    """Test case for list_flow

    
    """
    params = [('PageSize', 56),
                    ('Page', 56),
                    ('PageToken', 'page_token_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    response = await client.request(
        method='GET',
        path='/v2/Flows',
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

@pytest.mark.skip("application/x-www-form-urlencoded not supported by Connexion")
async def test_update_flow(client):
    """Test case for update_flow

    
    """
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/x-www-form-urlencoded',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    data = {
        'commit_message': 'commit_message_example',
        'definition': None,
        'friendly_name': 'friendly_name_example',
        'status': openapi_server.FlowEnumStatus()
        }
    response = await client.request(
        method='POST',
        path='/v2/Flows/{sid}'.format(sid='sid_example'),
        headers=headers,
        data=data,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

